Updated DeadActor() exception behavior.
DeadActor exception used to be raised when the weakref to
the actor would be scooped up by gc. The reference to the
Actor class became None in its Address class. Any access
to the actor class then would raise a DeadActor exception.

However, it seems that the latest eventlet hubs retain a
reference to actors' green thread indefinitely. So DeadActor
exception is never raised.

This commit tries to fix the DeadActor behavior by explicitly
checking actor.dead attribute during each access attempt in
addition to checking for None reference.

However, this has caused the 'kill' unit test to fail. There
might be some inconsistent behavior between when the Kill
exception should be raised and when the DeadActor exception
should be raised. For example the kill unit test does this:
  address.kill()
  address.wait()
  assertRaises(actor.Killed, actor.spawn(...).wait)

This used to work perhaps because the reference to the actor
was still there after it has been killed, so wait() would
raise actor.Killed even when called after the green thread
died. BUT if DeadActor is supposed to be raised when accessing
a .dead actor, then the call "address.wait()" would raise
a DeadActor because it is raised when the actor is referenced,
before even executing the wait() method of the event.

The test and _the semantics_ of the code has now changed such
that actor.Killed is only raised for anyone _already_ waiting.
If anyone starts waiting after the kill, they'll get a
DeadActor exception instead.

So the test was changed to:
  eventlet.spawn_after(1, lambda : address.kill())
  address.wait()
i.e. start waiting first then schedule a kill later.

Not sure if this is the correct behavior. Perhaps DeadActor exception
feature doesn't make sense at all?