Make shape.py correctly match JSON string types in Python 2.x
Python 2.x's 'json' module will decode a 'str' type as 'unicode'.
This is according to the JSON spec, which states that strings
are composed of 'any-Unicode-character...' (http://json.org).
So when a user specifies 'str' as the type to match it works
correctly in Python 3.x, where strings are unicode, but it
doesn't work in Python 2.x, which has both 'str' and 'unicode'
types.

This commit adds special code to cast all str types to unicode
in shape.py if Python's major version is 2.