Implemented convenient Actor.response() method for call responses.
Make it easier for a regular Actor (non-Server) to respond to calls
without having to deconstruct the whole call message to extract
messageid and address. Instead just call:
  self.receive(orig_message, response)

This method does the extraction of the address and message_id
behind the scenes and casts the response back.

Example:

   def main(self):
      pat,msg = self.receive(CALL_PATTERN)
      result = dispatch[msg['method']](msg['message'])
      self.respond(msg, result)

It is also possible to handle 2 types of errors in responses:
 1. Invalid method :
      self.respond_invalid_method(orig_message, method)
 2. Exception:
      self.respond_exception(orig_message, exception)

Also, made message optional for call and responses. By default
message and responses are None (JSON nulls).

Added self.respond* unittests.