Consolidate receive and receive_nowait by using a timeout.
receive(...) acquired a timeout _keyword_ argument.

If timeout is not specified, then the semantics of receive(...) has
not changed.

If timeout is 0, then the mailbox is checked only once and if nothing
is matched then (None,None) is returned. This is equivalent to the
old receive_nowait(...) call.

If timeout > 0 the mailbox is checked for timeout seconds and if nothing
matches (None,None) will be returned.

Added unittest for receive(...,timeout=0) and receive(...,timeout>0) cases.