Prevent greenlet.switch() to be scheduled multiple times.
A switch() to the actor is scheduled when a message send to
that actor. If multiple messages are sent multiple switches
would be scheduled. This could result in an extra switch,
for example, when actor's greenlet is waiting for IO.

It is necessary to only schedule a switch if the actor is
waiting in the receive loop and a switch already hasn't been
scheduled.